requirements.txt
transformers
torch
Pillow
gradio
